% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoboot-package.R
\docType{package}
\name{lassoboot-package}
\alias{lassoboot-package}
\alias{lassoboot}
\title{lassoboot}
\description{
Bootstrap methods for selecting the tuning parameter for LASSO-type regression models
}
\author{
replication files (maintainer) \email{replication.files@gmail.com}
}
\keyword{package}
